requirements.txt
autopep8==1.5.6
-e git+https://github.com/Fl0k3n/battle-ships.git@31955bec6ee9ecf1f1ee6ddd549446b73eea0c30#egg=battle_ships
pycodestyle==2.7.0
python-dotenv==0.17.0
toml==0.10.2
